Consolidate export helpers into file_io
